+++ src/Controllers/Const/Const.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace lovis.Controllers.Const
+{
+    public class Const
+    {
+        public static string EmailUsername = "lovisnotification@gmail.com";
+        public static string EmailPassword = "pelotudo1337";
+
+    }
+}