+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 
 [LICENSE](LICENSE) - [README](README.md) - [CHANGELOG](CHANGELOG.md) 
 
+## [0.3.2] - 01/06/2017 - [@emimontesdeoca](https://github.com/emimontesdeoca)
+### Add
+- Email confirmation sent when register.
+- User needs to confirm in order to actually register.
+
 ## [0.3.1] - 31/05/2017 - [@emimontesdeoca](https://github.com/emimontesdeoca)
 ### Add
 - Elements showing in project.